+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fixed: adding multiples of an item now works properly.
 
 ## [0.16.1] - 2018-11-25
 - Fixed: pressing the Reset button in the Item Quantity works properly.